requirements.txt
Flask==3.1.0
Pillow==11.1.0
python-dotenv==1.0.1
PyYAML==6.0.2
